requirements.txt
# Framework y ServidorDjango==5.2.5gunicorn==23.0.0# Base de Datosdj-database-url==3.0.1psycopg2-binary==2.9.10# Archivos Estáticoswhitenoise[brotli]==6.9.0# Utilidades y Herramientas de Djangodjango-otp==1.6.1qrcode==8.2django-widget-tweaks==1.5.0# Variables de Entorno y Peticiones HTTPpython-dotenv==1.0.1requests==2.32.5 